add event bus

+++ lambda/invoices/invoiceImportFunction.ts
@@ -1,5 +1,5 @@
 import { Context, S3Event, S3EventRecord } from "aws-lambda"
-import { ApiGatewayManagementApi, DynamoDB, S3 } from "aws-sdk"
+import { ApiGatewayManagementApi, DynamoDB, EventBridge, S3 } from "aws-sdk"
 import * as AWSXRay from "aws-xray-sdk"
 import { InvoiceTransactionRepository, InvoiceTransactionStatus } from "/opt/nodejs/invoiceTransaction"
 import { InvoiceWSService } from "/opt/nodejs/invoiceWSConnection"
@@ -9,15 +9,17 @@ AWSXRay.captureAWS(require('aws-sdk'))
 
 const invoicesDdb = process.env.INVOICE_DDB!
 const invoicesWsApiEndPoint = process.env.INVOICE_WSAPI_ENDPOINT!.substring(6)
+const auditBusName = process.env.AUDIT_BUS_NAME! 
 
 const s3Client = new S3()
 const ddbClient = new DynamoDB.DocumentClient()
 const apigwManagementApi = new ApiGatewayManagementApi({
   endpoint: invoicesWsApiEndPoint
 })
+const eventBridgeClient = new EventBridge() 
 
 const invoiceTransactionRepository = new InvoiceTransactionRepository(ddbClient, invoicesDdb)
-const invoiceWsService = new InvoiceWSService(apigwManagementApi)
+const invoiceWSService = new InvoiceWSService(apigwManagementApi)
 const invoiceRepository = new InvoiceRepository(ddbClient, invoicesDdb)
 
 export async function handler(event: S3Event, context: Context): Promise<void> {
@@ -39,11 +41,11 @@ async function processRecord(record: S3EventRecord) {
     const invoiceTransaction = await invoiceTransactionRepository.getInvoiceTransaction(key)
     if (invoiceTransaction.transactionStatus === InvoiceTransactionStatus.GENERATED){
       await Promise.all([
-        invoiceWsService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.RECEIVED),
+        invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.RECEIVED),
         invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.RECEIVED)
       ])
     } else {
-      await invoiceWsService.sendInvoiceStatus(
+      await invoiceWSService.sendInvoiceStatus(
         key, 
         invoiceTransaction.connectionId, 
         invoiceTransaction.transactionStatus
@@ -51,6 +53,7 @@ async function processRecord(record: S3EventRecord) {
       console.error("Non valid transaction status")
       return
     }
+
     const object = await s3Client.getObject({
       Bucket: record.s3.bucket.name,
       Key: key
@@ -59,24 +62,55 @@ async function processRecord(record: S3EventRecord) {
     const invoice = JSON.parse(object.Body!.toString('utf-8')) as InvoiceFile
     console.log(invoice)
 
-    await invoiceRepository.create({
-      pk: `#invoice_${invoice.customerName}`,
-      sk: invoice.invoiceNumber,
-      ttl: 0,
-      totalValue: invoice.totalValue,
-      productId: invoice.productId,
-      quantity: invoice.quantity,
-      transactionId: key,
-      createAt: Date.now()
-    })
+    if (invoice.invoiceNumber.length >= 5) {
+      const createInvoicePromise = await invoiceRepository.create({
+        pk: `#invoice_${invoice.customerName}`,
+        sk: invoice.invoiceNumber,
+        ttl: 0,
+        totalValue: invoice.totalValue,
+        productId: invoice.productId,
+        quantity: invoice.quantity,
+        transactionId: key,
+        createAt: Date.now()
+      })
 
-    await s3Client.deleteObject({
-      Bucket: record.s3.bucket.name,
-      Key: key
-    }).promise()
+      const deleteObjectPromise = s3Client.deleteObject({
+        Key: key,
+        Bucket: record.s3.bucket.name
+      }).promise()
 
-    await invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED)
-    await invoiceWsService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
+      const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, 
+        InvoiceTransactionStatus.PROCESSED)
+
+      const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, 
+        InvoiceTransactionStatus.PROCESSED)
+
+     await Promise.all([createInvoicePromise, deleteObjectPromise, updateInvoicePromise, sendStatusPromise])   
+  } else {
+    console.error(`Invoice import failed - non valid invoice number - TransactionId: ${key}`)
+
+      const putEventPromise = eventBridgeClient.putEvents({
+        Entries: [
+          {
+            Source: 'app.invoice',
+            EventBusName: auditBusName,
+            DetailType: 'invoice',
+            Time: new Date(),
+            Detail: JSON.stringify({
+              errorDetail: 'FAIL_NO_INVOICE_NUMBER',
+              info: {
+                  invoiceKey: key,
+                  customerName: invoice.customerName
+              }
+            })
+          }
+        ]
+      }).promise()
+
+      const sendStatusPromise = await invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED)
+      const updateInvoicePromise = await invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
+      await Promise.all([sendStatusPromise, updateInvoicePromise, putEventPromise])
+    }
   } catch (error) {
     console.log((<Error>error).message)
   }
